fix(assets): fix USD deposit button in coin intro
